// frontend/web/src/services/API/proserve-wb-packaging-api/models/MandatoryComponentsList.ts
export interface MandatoryComponentsList {
    /**
     * Platform for which the mandatory components list is enforced
     * @type {string}
     * @memberof MandatoryComponentsList
     */
    mandatoryComponentsListPlatform: string;
    /**
     * OS for which the mandatory components list is enforced
     * @type {string}
     * @memberof MandatoryComponentsList
     */
    mandatoryComponentsListOsVersion: string;
    /**
     * Architecture for which the mandatory components list is enforced
     * @type {string}
     * @memberof MandatoryComponentsList
     */
    mandatoryComponentsListArchitecture: string;
    /**
     * Components versions that will be prepended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof MandatoryComponentsList
     */
    prependedComponentsVersions: Array<ComponentVersionEntry>;
    /**
     * Components versions that will be appended to recipe versions
     * @type {Array<ComponentVersionEntry>}
     * @memberof MandatoryComponentsList
     */
    appendedComponentsVersions: Array<ComponentVersionEntry>;
}